修改SSM项目 12.26  02

+++ heima_ssm/heima_ssm_service/src/main/java/com/itheima/service/impl/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.itheima.service.impl;
 
+import com.github.pagehelper.PageHelper;
 import com.itheima.dao.ProductDao;
 import com.itheima.domain.Product;
 import com.itheima.service.ProductService;
@@ -12,17 +13,32 @@ public class ProductServiceImpl implements ProductService {
     @Autowired
     private ProductDao productDao;
     @Override
-    public List<Product> findAll() {
-        return productDao.findAll();
+    public List<Product> findAll(String str,int pageNum,int pageSize) {
+        PageHelper.startPage(pageNum,pageSize);
+        return productDao.findAll(str);
     }
 
     @Override
-    public void delete(String id) {
-        productDao.delete(id);
+    public Product findById(String id) {
+        return productDao.findById(id);
+    }
+
+    @Override
+    public void delete(String[] ids) {
+        if(ids!=null&&ids.length>0){
+            for (String id : ids) {
+                productDao.delete(id);
+            }
+        }
     }
 
     @Override
     public void save(Product product) {
         productDao.save(product);
+    }
+
+    @Override
+    public void update(Product product) {
+        productDao.update(product);
     }
 }
